Resolved all TODOs. Added verification code expiration logic

+++ src/main/java/com/serhiihurin/shop/online_shop/facades/UserFacadeImpl.java
@@ -6,15 +6,19 @@ import com.serhiihurin.shop.online_shop.dto.UserRequestDTO;
 import com.serhiihurin.shop.online_shop.dto.UsernameUpdateResponseDTO;
 import com.serhiihurin.shop.online_shop.entity.User;
 import com.serhiihurin.shop.online_shop.dto.RegisterRequestDTO;
+import com.serhiihurin.shop.online_shop.entity.VerificationCode;
 import com.serhiihurin.shop.online_shop.exception.ApiRequestException;
+import com.serhiihurin.shop.online_shop.exception.VerificationCodeExpirationException;
 import com.serhiihurin.shop.online_shop.services.UserService;
 import com.serhiihurin.shop.online_shop.services.JWTService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Component
@@ -25,6 +29,8 @@ public class UserFacadeImpl implements UserFacade {
     private final JWTService jwtService;
     private final ModelMapper modelMapper;
     private final VerificationCodeRepository verificationCodeRepository;
+    @Value("${custom.verification-code-expiration}")
+    private int expiration;
 
     @Override
     public List<User> getAllUsers() {
@@ -73,11 +79,15 @@ public class UserFacadeImpl implements UserFacade {
     @Transactional
     @Override
     public void updatePassword(User currentAuthenticatedUser, PasswordUpdateRequestDTO passwordUpdateRequestDTO) {
-        //TODO nullpointer?
-        String verificationCode =
-                verificationCodeRepository.findByUserId(currentAuthenticatedUser.getId()).getVerificationCode();
-        if (!verificationCode.equals(passwordUpdateRequestDTO.getVerificationCode())) {
+        //TODO nullpointer? //done
+        VerificationCode verificationCode =
+                verificationCodeRepository.findByUserId(currentAuthenticatedUser.getId())
+                        .orElseThrow(() -> new ApiRequestException("Missing verification code"));
+        if (!verificationCode.getVerificationCode().equals(passwordUpdateRequestDTO.getVerificationCode())) {
             throw new ApiRequestException("Invalid verification code. Try again");
+        }
+        if (LocalDateTime.now().isAfter(verificationCode.getCreationTime().plusMinutes(expiration))) {
+            throw new VerificationCodeExpirationException("Verification code is no longer valid");
         }
         verificationCodeRepository.deleteVerificationCodeByUserId(currentAuthenticatedUser.getId());
         userService.updatePassword(currentAuthenticatedUser,passwordUpdateRequestDTO);
@@ -86,6 +96,9 @@ public class UserFacadeImpl implements UserFacade {
 
     @Override
     public void deleteUser(Long id) {
+        if (id == null) {
+            throw new ApiRequestException("Invalid URL. Parameter id must not be null");
+        }
         userService.deleteUser(id);
     }
 }
